+++ src/TestAsync/Program.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using GetSomeInput;
 using RestWrapper;
 
 namespace TestAsync
@@ -48,14 +49,14 @@ namespace TestAsync
 
                         case "put":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Put,
                                 null,
-                                InputString("Content type:", "text/plain", false));
+                                Inputty.GetString("Content type:", "text/plain", false));
 
                             if (debug) req.Logger = Console.WriteLine;
 
-                            data = Encoding.UTF8.GetBytes(InputString("Data:", "Hello, world!", false));
+                            data = Encoding.UTF8.GetBytes(Inputty.GetString("Data:", "Hello, world!", false));
                             resp = req.SendAsync(data).Result;
                             if (resp == null)
                             {
@@ -70,14 +71,14 @@ namespace TestAsync
 
                         case "post":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Post,
                                 null,
-                                InputString("Content type:", "text/plain", false));
+                                Inputty.GetString("Content type:", "text/plain", false));
 
                             if (debug) req.Logger = Console.WriteLine;
 
-                            data = Encoding.UTF8.GetBytes(InputString("Data:", "Hello, world!", false));
+                            data = Encoding.UTF8.GetBytes(Inputty.GetString("Data:", "Hello, world!", false));
                             resp = req.SendAsync(data).Result;
                             if (resp == null)
                             {
@@ -92,14 +93,14 @@ namespace TestAsync
 
                         case "delete":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Delete,
                                 null,
-                                InputString("Content type:", "text/plain", false));
+                                Inputty.GetString("Content type:", "text/plain", false));
 
                             if (debug) req.Logger = Console.WriteLine;
 
-                            data = Encoding.UTF8.GetBytes(InputString("Data:", "Hello, world!", false));
+                            data = Encoding.UTF8.GetBytes(Inputty.GetString("Data:", "Hello, world!", false));
                             resp = req.SendAsync(data).Result;
                             if (resp == null)
                             {
@@ -114,7 +115,7 @@ namespace TestAsync
 
                         case "head":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Head,
                                 null,
                                 null);
@@ -135,7 +136,7 @@ namespace TestAsync
 
                         case "get":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Get,
                                 null,
                                 null);
@@ -151,6 +152,20 @@ namespace TestAsync
                             {
                                 Console.WriteLine(resp.ToString());
                                 Console.WriteLine(resp.DataAsString);
+                            }
+                            break;
+
+                        case "query":
+                            req = new RestRequest(
+                                Inputty.GetString("URL:", "http://localhost:8888/?foo&bar=baz&another=value&key=val", false),
+                                HttpMethod.Get);
+
+                            if (req.Query.AllKeys.Count() > 0)
+                            {
+                                for (int i = 0; i < req.Query.AllKeys.Count(); i++)
+                                {
+                                    Console.WriteLine(req.Query.AllKeys[i] + ": " + req.Query.Get(i));
+                                }
                             }
                             break;
 
@@ -180,7 +195,9 @@ namespace TestAsync
             Console.WriteLine("  post        Submit a POST request");
             Console.WriteLine("  delete      Submit a DELETE request");
             Console.WriteLine("  head        Submit a HEAD request");
+            Console.WriteLine("  query       Retrieve the query from a given URL");
             Console.WriteLine("  debug       Enable or disable console debugging (currently " + debug + ")");
+            Console.WriteLine("");
         }
 
         static string StackToString()
@@ -227,117 +244,6 @@ namespace TestAsync
             Console.WriteLine("---");
 
             return;
-        }
-
-        static string InputString(string question, string defaultAnswer, bool allowNull)
-        {
-            while (true)
-            {
-                Console.Write(question);
-
-                if (!String.IsNullOrEmpty(defaultAnswer))
-                {
-                    Console.Write(" [" + defaultAnswer + "]");
-                }
-
-                Console.Write(" ");
-
-                string userInput = Console.ReadLine();
-
-                if (String.IsNullOrEmpty(userInput))
-                {
-                    if (!String.IsNullOrEmpty(defaultAnswer)) return defaultAnswer;
-                    if (allowNull) return null;
-                    else continue;
-                }
-
-                return userInput;
-            }
-        }
-
-        static int InputInteger(string question, int defaultAnswer, bool positiveOnly, bool allowZero)
-        {
-            while (true)
-            {
-                Console.Write(question);
-                Console.Write(" [" + defaultAnswer + "] ");
-
-                string userInput = Console.ReadLine();
-
-                if (String.IsNullOrEmpty(userInput))
-                {
-                    return defaultAnswer;
-                }
-
-                int ret = 0;
-                if (!Int32.TryParse(userInput, out ret))
-                {
-                    Console.WriteLine("Please enter a valid integer.");
-                    continue;
-                }
-
-                if (ret == 0)
-                {
-                    if (allowZero)
-                    {
-                        return 0;
-                    }
-                }
-
-                if (ret < 0)
-                {
-                    if (positiveOnly)
-                    {
-                        Console.WriteLine("Please enter a value greater than zero.");
-                        continue;
-                    }
-                }
-
-                return ret;
-            }
-        }
-
-        static bool InputBoolean(string question, bool trueDefault)
-        {
-            Console.Write(question);
-
-            if (trueDefault) Console.Write(" [Y/n]? ");
-            else Console.Write(" [y/N]? ");
-
-            string userInput = Console.ReadLine();
-
-            if (String.IsNullOrEmpty(userInput))
-            {
-                if (trueDefault) return true;
-                return false;
-            }
-
-            userInput = userInput.ToLower();
-
-            if (trueDefault)
-            {
-                if (
-                    (String.Compare(userInput, "n") == 0)
-                    || (String.Compare(userInput, "no") == 0)
-                   )
-                {
-                    return false;
-                }
-
-                return true;
-            }
-            else
-            {
-                if (
-                    (String.Compare(userInput, "y") == 0)
-                    || (String.Compare(userInput, "yes") == 0)
-                   )
-                {
-                    return true;
-                }
-
-                return false;
-            }
         }
     }
 }