+++ src/TestStream/Program.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using GetSomeInput;
 using RestWrapper;
 
 namespace TestStream
@@ -50,14 +51,14 @@ namespace TestStream
 
                         case "put":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Put,
                                 null,
-                                InputString("Content type:", "text/plain", false));
+                                Inputty.GetString("Content type:", "text/plain", false));
 
                             if (debug) req.Logger = Console.WriteLine;
 
-                            data = Encoding.UTF8.GetBytes(InputString("Data:", "Hello, world!", false));
+                            data = Encoding.UTF8.GetBytes(Inputty.GetString("Data:", "Hello, world!", false));
                             ms = new MemoryStream(data);
                             resp = req.Send(data.Length, ms);
                             if (resp == null)
@@ -77,14 +78,14 @@ namespace TestStream
 
                         case "post":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Post,
                                 null,
-                                InputString("Content type:", "text/plain", false));
+                                Inputty.GetString("Content type:", "text/plain", false));
 
                             if (debug) req.Logger = Console.WriteLine;
 
-                            data = Encoding.UTF8.GetBytes(InputString("Data:", "Hello, world!", false));
+                            data = Encoding.UTF8.GetBytes(Inputty.GetString("Data:", "Hello, world!", false));
                             ms = new MemoryStream(data);
                             resp = req.Send(data.Length, ms);
                             if (resp == null)
@@ -104,14 +105,14 @@ namespace TestStream
 
                         case "delete":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Delete,
                                 null,
-                                InputString("Content type:", "text/plain", false));
+                                Inputty.GetString("Content type:", "text/plain", false));
 
                             if (debug) req.Logger = Console.WriteLine;
 
-                            data = Encoding.UTF8.GetBytes(InputString("Data:", "Hello, world!", false));
+                            data = Encoding.UTF8.GetBytes(Inputty.GetString("Data:", "Hello, world!", false));
                             ms = new MemoryStream(data);
                             resp = req.Send(data.Length, ms);
                             if (resp == null)
@@ -131,7 +132,7 @@ namespace TestStream
 
                         case "head":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Head,
                                 null,
                                 null);
@@ -156,7 +157,7 @@ namespace TestStream
 
                         case "get":
                             req = new RestRequest(
-                                InputString("URL:", "http://127.0.0.1:8000/", false),
+                                Inputty.GetString("URL:", "http://127.0.0.1:8000/", false),
                                 HttpMethod.Get,
                                 null,
                                 null);
@@ -175,6 +176,20 @@ namespace TestStream
                                 {
                                     data = StreamToBytes(resp.Data);
                                     Console.WriteLine(Encoding.UTF8.GetString(data));
+                                }
+                            }
+                            break;
+
+                        case "query":
+                            req = new RestRequest(
+                                Inputty.GetString("URL:", "http://localhost:8888/?foo&bar=baz&another=value&key=val", false),
+                                HttpMethod.Get);
+
+                            if (req.Query.AllKeys.Count() > 0)
+                            {
+                                for (int i = 0; i < req.Query.AllKeys.Count(); i++)
+                                {
+                                    Console.WriteLine(req.Query.AllKeys[i] + ": " + req.Query.Get(i));
                                 }
                             }
                             break;
@@ -205,7 +220,9 @@ namespace TestStream
             Console.WriteLine("  post        Submit a POST request");
             Console.WriteLine("  delete      Submit a DELETE request");
             Console.WriteLine("  head        Submit a HEAD request");
+            Console.WriteLine("  query       Retrieve the query from a given URL");
             Console.WriteLine("  debug       Enable or disable console debugging (currently " + debug + ")");
+            Console.WriteLine("");
         }
 
         static string StackToString()
@@ -252,117 +269,6 @@ namespace TestStream
             Console.WriteLine("---");
 
             return;
-        }
-
-        static string InputString(string question, string defaultAnswer, bool allowNull)
-        {
-            while (true)
-            {
-                Console.Write(question);
-
-                if (!String.IsNullOrEmpty(defaultAnswer))
-                {
-                    Console.Write(" [" + defaultAnswer + "]");
-                }
-
-                Console.Write(" ");
-
-                string userInput = Console.ReadLine();
-
-                if (String.IsNullOrEmpty(userInput))
-                {
-                    if (!String.IsNullOrEmpty(defaultAnswer)) return defaultAnswer;
-                    if (allowNull) return null;
-                    else continue;
-                }
-
-                return userInput;
-            }
-        }
-
-        static int InputInteger(string question, int defaultAnswer, bool positiveOnly, bool allowZero)
-        {
-            while (true)
-            {
-                Console.Write(question);
-                Console.Write(" [" + defaultAnswer + "] ");
-
-                string userInput = Console.ReadLine();
-
-                if (String.IsNullOrEmpty(userInput))
-                {
-                    return defaultAnswer;
-                }
-
-                int ret = 0;
-                if (!Int32.TryParse(userInput, out ret))
-                {
-                    Console.WriteLine("Please enter a valid integer.");
-                    continue;
-                }
-
-                if (ret == 0)
-                {
-                    if (allowZero)
-                    {
-                        return 0;
-                    }
-                }
-
-                if (ret < 0)
-                {
-                    if (positiveOnly)
-                    {
-                        Console.WriteLine("Please enter a value greater than zero.");
-                        continue;
-                    }
-                }
-
-                return ret;
-            }
-        }
-
-        static bool InputBoolean(string question, bool trueDefault)
-        {
-            Console.Write(question);
-
-            if (trueDefault) Console.Write(" [Y/n]? ");
-            else Console.Write(" [y/N]? ");
-
-            string userInput = Console.ReadLine();
-
-            if (String.IsNullOrEmpty(userInput))
-            {
-                if (trueDefault) return true;
-                return false;
-            }
-
-            userInput = userInput.ToLower();
-
-            if (trueDefault)
-            {
-                if (
-                    (String.Compare(userInput, "n") == 0)
-                    || (String.Compare(userInput, "no") == 0)
-                   )
-                {
-                    return false;
-                }
-
-                return true;
-            }
-            else
-            {
-                if (
-                    (String.Compare(userInput, "y") == 0)
-                    || (String.Compare(userInput, "yes") == 0)
-                   )
-                {
-                    return true;
-                }
-
-                return false;
-            }
         }
 
         static byte[] StreamToBytes(Stream input)